src/docker_image/video-encoding/requirements_2.txt
numpy==1.16.6
Pillow==3.1.2
scipy==1.2.3
kafka-python==2.0.1
